Move commitOutputdir to FileStorageManager.
